+++ packages/plugin-admin-api/test/legacy/alerts-api.spec.js
@@ -96,8 +96,8 @@ describe("alerts api", () => {
 
     it("should return alert containing error", async () => {
       const response = await doLegacyFetch("/admin/alerts");
-      expect(response.body.length).toEqual(6);
-      expect(response.body[5].error.message).toEqual(
+      expect(response.body.length).toEqual(7);
+      expect(response.body[6].error.message).toEqual(
         expect.stringContaining("Cannot find module '../db/users'")
       );
     });