+++ test/plugin-inquirer-cli-e2e/src/main/files-watch.spec.js
@@ -15,7 +15,6 @@ const {
   waitForServerAndCli,
   wait,
   fixturesFolder,
-  pathJoin,
 } = require("./support/helpers");
 
 describe("files watcher", () => {
@@ -161,19 +160,12 @@ describe("files watcher", () => {
     });
 
     it("should display an error", async () => {
+      expect(mocks.currentScreen).toEqual(expect.stringContaining("Error: [files:load"));
+      expect(mocks.currentScreen).toEqual(expect.stringContaining("Error loading file"));
       expect(mocks.currentScreen).toEqual(
-        expect.stringContaining(
-          "Error: [files:collections:error] Error loading collections from file"
-        )
+        expect.stringContaining(`collections.js: foo is not defined`)
       );
-      expect(mocks.currentScreen).toEqual(
-        expect.stringContaining(
-          `${pathJoin("main", "fixtures", "temp", "collections.js")}: foo is not defined`
-        )
-      );
-      expect(mocks.currentScreen).toEqual(
-        expect.stringContaining(`${pathJoin("main", "fixtures", "temp", "collections.js")}:11:18`)
-      );
+      expect(mocks.currentScreen).toEqual(expect.stringContaining(`collections.js:11:18`));
     });
 
     it("should have no mocks available", async () => {
@@ -212,20 +204,17 @@ describe("files watcher", () => {
       expect.assertions(2);
       expect(mocks.currentScreen).toEqual(expect.not.stringContaining("ALERTS"));
       expect(mocks.currentScreen).toEqual(
-        expect.stringContaining("Error loading collections from file")
+        expect.stringContaining("[error][alerts:load] Error loading file")
       );
     });
 
     it("should display alerts when exit logs mode", async () => {
-      expect.assertions(3);
+      expect.assertions(4);
       await mocks.pressEnter();
       await wait(2000);
       expect(mocks.currentScreen).toEqual(expect.not.stringContaining("Displaying logs"));
-      expect(mocks.currentScreen).toEqual(
-        expect.stringContaining(
-          "Error: [files:collections:error] Error loading collections from file"
-        )
-      );
+      expect(mocks.currentScreen).toEqual(expect.stringContaining("Error: [files:load"));
+      expect(mocks.currentScreen).toEqual(expect.stringContaining("Error loading file"));
       expect(mocks.currentScreen).toEqual(expect.stringContaining("ALERTS"));
     });